web/requirements.txt
flask==0.10.1
markdown==2.3.1
pyyaml==3.10
waitress==0.8.5
